refactor: change the structure in src folder
- Add utilities folder to contain the utilities offered by the plugin
- Add utils folder
- Change suffix in the utilities files to utility
