+++ tsconfig.json
@@ -108,5 +108,5 @@
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */,
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["app"]
+  "exclude": ["app", "node_modules", "dist"]
 }
